backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0
pymongo==4.7.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
mediapipe==0.10.14
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.3.0
websockets==12.0
httpx==0.27.0
aiofiles==23.2.1
